Move old Publication API calls to separate file
