+++ lib/bal/api.js
@@ -1,4 +1,4 @@
-export const BACKEND_URL = process.env.NEXT_PUBLIC_BACKEND_URL || 'https://backend.adresse.data.gouv.fr'
+const BACKEND_URL = process.env.NEXT_PUBLIC_BACKEND_URL || 'https://backend.adresse.data.gouv.fr'
 
 async function _fetch(url, method, body) {
   const options = {
@@ -26,83 +26,6 @@ async function _fetch(url, method, body) {
   }
 
   throw new Error('Une erreur est survenue')
-}
-
-export async function extractCommunes(communes) {
-  const url = `${BACKEND_URL}/ban/extract?communes=${communes.map(c => c.code).join()}`
-  const options = {
-    headers: {
-      Accept: 'text/csv',
-      'Access-Control-Request-Headers': 'Content-Type, Content-Disposition'
-    },
-    mode: 'cors',
-    method: 'GET'
-  }
-
-  const response = await fetch(url, options)
-  const blob = await response.blob()
-  return blob
-}
-
-export async function uploadCSV(file) {
-  const href = `${BACKEND_URL}/publication/submissions/upload`
-  const options = {
-    method: 'POST',
-    mode: 'cors',
-    headers: {'Content-Type': 'text/csv'},
-    body: file
-  }
-  const response = await fetch(href, options)
-
-  if (!response.ok) {
-    const {message} = await response.json()
-    throw new Error(message)
-  }
-
-  if (response.ok) {
-    return response.json()
-  }
-
-  throw new Error('Impossible d’envoyer le fichier CSV, veuillez réessayer ultérieurement.')
-}
-
-export async function submissionsBal(url) {
-  const href = `${BACKEND_URL}/publication/submissions`
-  const response = await _fetch(href, 'POST', {url})
-
-  return response
-}
-
-export async function getSubmissions(submissionId) {
-  const href = `${BACKEND_URL}/publication/submissions/${submissionId}`
-  const response = await _fetch(href, 'GET')
-
-  return response
-}
-
-export async function askAuthentificationCode(submissionId) {
-  const href = `${BACKEND_URL}/publication/submissions/${submissionId}/send-pin-code`
-  const options = {
-    method: 'POST',
-    mode: 'cors',
-    headers: {'Content-Type': 'application/json'}
-  }
-
-  return fetch(href, options)
-}
-
-export async function submitAuthentificationCode(submissionId, code) {
-  const href = `${BACKEND_URL}/publication/submissions/${submissionId}/validate-pin-code`
-  const response = await _fetch(href, 'POST', {code})
-
-  return response
-}
-
-export async function submitBal(submissionId) {
-  const href = `${BACKEND_URL}/publication/submissions/${submissionId}/publish`
-  const response = await _fetch(href, 'POST', {})
-
-  return response
 }
 
 export async function getDatasets() {
